+++ apps/client/src/app/pages/gearset/materias-needed-popup/materias-needed-popup.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { MateriaService } from '../../../modules/gearsets/materia.service';
 import { TeamcraftGearset } from '../../../model/gearset/teamcraft-gearset';
 
@@ -12,13 +12,24 @@ export class MateriasNeededPopupComponent implements OnInit {
   @Input()
   gearset: TeamcraftGearset;
 
+  @Input()
+  includeAllTools = false;
+
+  @Output()
+  includeAllToolsChange = new EventEmitter<boolean>();
+
   totalNeeded: { id: number, amount: number }[] = [];
 
   constructor(private materiaService: MateriaService) {
   }
 
+  computeRequirements(): void {
+    localStorage.setItem('gearsets:include-all-tools', this.includeAllTools.toString());
+    this.totalNeeded = this.materiaService.getTotalNeededMaterias(this.gearset, this.includeAllTools);
+  }
+
   ngOnInit(): void {
-    this.totalNeeded = this.materiaService.getTotalNeededMaterias(this.gearset);
+    this.computeRequirements();
   }
 
 }