+++ data-extraction/xivapi/index.js
@@ -1818,7 +1818,9 @@ if (hasTodo('foods')) {
   const foods = [];
   getAllPages('https://xivapi.com/Search?indexes=items&filters=ItemSearchCategory.ID=45&columns=ID,Bonuses,LevelItem,LevelEquip').subscribe(page => {
     page.Results.forEach(entry => {
-      foods.push(entry);
+      if (entry.Bonuses) {
+        foods.push(entry);
+      }
     });
   }, null, () => {
     persistToJsonAsset('foods', foods);